+++ api/src/db/migrations/2022-03-16-migrating.js
@@ -1,0 +1,7 @@
+const sequelize = require("../sequelize");
+
+sequelize.query(`
+  ALTER TABLE "mano"."Organisation"
+  ADD COLUMN IF NOT EXISTS "migrating" boolean DEFAULT false,
+  ADD COLUMN IF NOT EXISTS "migrationLastUpdateAt" timestamp with time zone;
+`);