+++ api/src/models/passage.js
@@ -1,0 +1,16 @@
+const { DataTypes, Model, Sequelize, Deferrable } = require("sequelize");
+const sequelize = require("../db/sequelize");
+
+class Passage extends Model {}
+
+const schema = {
+  _id: { type: DataTypes.UUID, allowNull: false, defaultValue: Sequelize.UUIDV4, primaryKey: true },
+  organisation: { type: DataTypes.UUID, references: { model: "Organisation", key: "_id", deferrable: Deferrable.INITIALLY_IMMEDIATE } },
+
+  encrypted: { type: DataTypes.TEXT },
+  encryptedEntityKey: { type: DataTypes.TEXT },
+};
+
+Passage.init(schema, { sequelize, modelName: "Passage", freezeTableName: true });
+
+module.exports = Passage;