+++ dashboard/src/components/Comments.js
@@ -1,7 +1,6 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import React, { useEffect, useMemo, useState } from 'react';
 import styled from 'styled-components';
-import dayjs from 'dayjs';
 import { Modal, Input, Button as CloseButton, Col, Row, ModalHeader, ModalBody, FormGroup, Label } from 'reactstrap';
 import { toastr } from 'react-redux-toastr';
 import DatePicker from 'react-datepicker';
@@ -20,8 +19,9 @@ import { useRecoilState, useRecoilValue } from 'recoil';
 import { dateForDatePicker } from '../services/date';
 import { loadingState } from './Loader';
 import useApi from '../services/api';
+import { formatDateTimeWithNameOfDay } from '../services/date';
 
-const Comments = ({ personId = '', actionId = '', forPassages = false, onUpdateResults }) => {
+const Comments = ({ personId = '', actionId = '', onUpdateResults }) => {
   const [editingId, setEditing] = useState(null);
   const [clearNewCommentKey, setClearNewCommentKey] = useState(null);
   const API = useApi();
@@ -34,18 +34,12 @@ const Comments = ({ personId = '', actionId = '', forPassages = false, onUpdateR
 
   const comments = useMemo(
     () =>
-      allComments
-        .filter((c) => {
-          if (!!personId) return c.person === personId;
-          if (!!actionId) return c.action === actionId;
-          return false;
-        })
-        .filter((c) => {
-          const commentIsPassage = c?.comment?.includes('Passage enregistré');
-          if (forPassages) return commentIsPassage;
-          return !commentIsPassage;
-        }),
-    [personId, actionId, forPassages, allComments]
+      allComments.filter((c) => {
+        if (!!personId) return c.person === personId;
+        if (!!actionId) return c.action === actionId;
+        return false;
+      }),
+    [personId, actionId, allComments]
   );
 
   useEffect(() => {
@@ -110,7 +104,7 @@ const Comments = ({ personId = '', actionId = '', forPassages = false, onUpdateR
     <React.Fragment>
       <Row style={{ marginTop: '30px', marginBottom: '5px' }}>
         <Col md={4}>
-          <Title>{!forPassages ? 'Commentaires' : 'Passages'}</Title>
+          <Title>Commentaires</Title>
         </Col>
       </Row>
       <Box>
@@ -118,14 +112,14 @@ const Comments = ({ personId = '', actionId = '', forPassages = false, onUpdateR
           <Loading />
         ) : (
           <>
-            <EditingComment key={clearNewCommentKey} onSubmit={addData} newComment forPassages={forPassages} />
+            <EditingComment key={clearNewCommentKey} onSubmit={addData} newComment />
             {comments.map((comment) => {
               return (
                 <StyledComment key={comment._id}>
                   <CloseButton close onClick={() => deleteData(comment._id)} />
                   <UserName id={comment.user} wrapper={(name) => <div className="author">{name}</div>} />
                   <div className="user"></div>
-                  <div className="time">{dayjs(comment.date || comment.createdAt).format('MMM DD, YYYY | hh:mm A')}</div>
+                  <div className="time">{formatDateTimeWithNameOfDay(comment.date || comment.createdAt)}</div>
                   <div className="content">
                     <p onClick={() => setEditing(comment._id)}>
                       {comment.comment
@@ -153,13 +147,12 @@ const Comments = ({ personId = '', actionId = '', forPassages = false, onUpdateR
         value={comments.find((c) => c._id === editingId)}
         onSubmit={updateData}
         onCancel={() => setEditing(null)}
-        forPassages={forPassages}
       />
     </React.Fragment>
   );
 };
 
-const EditingComment = ({ value = {}, commentId, onSubmit, onCancel, newComment, forPassages }) => {
+const EditingComment = ({ value = {}, commentId, onSubmit, onCancel, newComment }) => {
   const user = useRecoilValue(userState);
   const [open, setOpen] = useState(false);
 
@@ -174,7 +167,7 @@ const EditingComment = ({ value = {}, commentId, onSubmit, onCancel, newComment,
 
   return (
     <>
-      {!!newComment && !forPassages && <ButtonCustom title="Ajouter un commentaire" onClick={() => setOpen(true)} style={{ marginBottom: 20 }} />}
+      {!!newComment && <ButtonCustom title="Ajouter un commentaire" onClick={() => setOpen(true)} style={{ marginBottom: 20 }} />}
       <Modal isOpen={!!open} toggle={onCancelRequest} size="lg">
         <ModalHeader toggle={onCancelRequest}>{newComment ? 'Créer un' : 'Éditer le'} commentaire</ModalHeader>
         <ModalBody>