+++ dashboard/src/scenes/stats/index.js
@@ -27,20 +27,23 @@ import { reportsState } from '../../recoil/reports';
 import ExportData from '../data-import-export/ExportData';
 import SelectCustom from '../../components/SelectCustom';
 import { territoriesState } from '../../recoil/territory';
-import { passagesNonAnonymousPerDatePerTeamSelector } from '../../recoil/selectors';
-import { dayjsInstance } from '../../services/date';
+import { getIsDayWithinHoursOffsetOfPeriod } from '../../services/date';
 import { loadingState, refreshTriggerState } from '../../components/Loader';
+import { passagesState } from '../../recoil/passages';
 
-const getDataForPeriod = (data, { startDate, endDate }, filters = []) => {
+const getDataForPeriod = (data, { startDate, endDate }, currentTeam, viewAllOrganisationData, { filters = [], field = 'createdAt' } = {}) => {
   if (!!filters?.filter((f) => Boolean(f?.value)).length) data = filterData(data, filters);
   if (!startDate || !endDate) {
     return data;
   }
-  return data.filter((item) => dayjsInstance(item.createdAt).isBetween(startDate, endDate));
+  const offsetHours = !!viewAllOrganisationData ? 0 : currentTeam?.nightSession ? 12 : 0;
+
+  return data.filter((item) =>
+    getIsDayWithinHoursOffsetOfPeriod(item[field] || item.createdAt, { referenceStartDay: startDate, referenceEndDay: endDate }, offsetHours)
+  );
 };
 
 const tabs = ['Général', 'Accueil', 'Actions', 'Personnes suivies', 'Observations', 'Comptes-rendus'];
-
 const Stats = () => {
   const organisation = useRecoilValue(organisationState);
   const user = useRecoilValue(userState);
@@ -51,6 +54,7 @@ const Stats = () => {
   const allActions = useRecoilValue(actionsState);
   const allreports = useRecoilValue(reportsState);
   const allObservations = useRecoilValue(territoryObservationsState);
+  const allPassages = useRecoilValue(passagesState);
   const customFieldsObs = useRecoilValue(customFieldsObsSelector);
   const customFieldsPersonsSocial = useRecoilValue(customFieldsPersonsSocialSelector);
   const customFieldsPersonsMedical = useRecoilValue(customFieldsPersonsMedicalSelector);
@@ -62,15 +66,6 @@ const Stats = () => {
   const [filterPersons, setFilterPersons] = useState([]);
   const [viewAllOrganisationData, setViewAllOrganisationData] = useState(teams.length === 1);
   const [period, setPeriod] = useState({ startDate: null, endDate: null });
-  const nonAnonymousPassages = useRecoilValue(
-    passagesNonAnonymousPerDatePerTeamSelector({
-      filterCurrentTeam: !viewAllOrganisationData,
-      date: {
-        startDate: period.startDate,
-        endDate: period.endDate,
-      },
-    })
-  );
 
   const addFilter = ({ field, value }) => {
     setFilterPersons((filters) => [...filters, { field, value }]);
@@ -81,22 +76,41 @@ const Stats = () => {
   const persons = getDataForPeriod(
     allPersons.filter((e) => viewAllOrganisationData || (e.assignedTeams || []).includes(currentTeam._id)),
     period,
-    filterPersons
+    currentTeam,
+    viewAllOrganisationData,
+    { filters: filterPersons }
   );
   const actions = getDataForPeriod(
     allActions.filter((e) => viewAllOrganisationData || e.team === currentTeam._id),
-    period
+    period,
+    currentTeam,
+    viewAllOrganisationData
   );
   const observations = getDataForPeriod(
     allObservations
       .filter((e) => viewAllOrganisationData || e.team === currentTeam._id)
       .filter((e) => !territory?._id || e.territory === territory._id),
-    period
+    period,
+    currentTeam,
+    viewAllOrganisationData,
+    { field: 'observedAt' }
   );
+  const passages = getDataForPeriod(
+    allPassages.filter((e) => viewAllOrganisationData || e.team === currentTeam._id),
+    period,
+    currentTeam,
+    viewAllOrganisationData,
+    { field: 'date' }
+  );
+
   const reports = getDataForPeriod(
     allreports.filter((e) => viewAllOrganisationData || e.team === currentTeam._id),
-    period
+    period,
+    currentTeam,
+    viewAllOrganisationData,
+    { field: 'date' }
   );
+
   const reportsServices = reports.map((rep) => (rep.services ? JSON.parse(rep.services) : null)).filter(Boolean);
 
   // Add enabled custom fields in filters.
@@ -176,10 +190,7 @@ const Stats = () => {
           <TabPane tabId={1}>
             <Title>Statistiques de l'accueil</Title>
             <Row>
-              <Block
-                data={reports.reduce((passages, rep) => passages + (rep.passages || 0), 0) + (nonAnonymousPassages?.length || 0)}
-                title="Nombre de passages"
-              />
+              <Block data={passages.length} title="Nombre de passages" />
               {organisation.services?.map((service) => (
                 <Block
                   key={service}